+++ src/utils/goal/regexGoal.ts
